# ci: pin xharness