@Flash0ver 2025-07-11
Still failing, but for a different reason:
**before this PR**
```Text
...
Failed to install the application
...
Failed to uninstall the app bundle! Check logs for more details!
XHarness exit code: 78 (PACKAGE_INSTALLATION_FAILURE)
...
```

**after this PR**
```Text
...
Application 'Sentry.Maui.Device.TestApp' was installed successfully on 'iPhone Xs (iOS 18.5) - created by XHarness'
...
Application finished the test run successfully with some failed tests
...
Application 'io.sentry.dotnet.maui.device.testapp' was uninstalled successfully
...
```

The test `Sentry.Tests.Internals.SentryStopwatchTests.Elapsed_IsValid` is failing
```Text
Expected time to be within 500ms from 100ms, but found 675ms and 212.8µs.
...
```

Seems to be a timing-related issue via `Thread.Sleep`.